auth router built, index refactored
